refactor: pull common functions out into utils
