+++ scripts/build-netlify.js
@@ -1,40 +1,10 @@
 const CopyWebpackPlugin = require('copy-webpack-plugin');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
 const MomentLocalesPlugin = require('moment-locales-webpack-plugin');
 const PostCompile = require('post-compile-webpack-plugin');
-const { author, name } = require('../package.json');
-const { handleWithPrefix, compressJSON, initMappingFiles } = require('./utils');
-
-const getCopyPluginProps = mappings => {
-  const assetsFiles = [
-    {
-      from: handleWithPrefix('src/assets'),
-      to: handleWithPrefix('dist/assets'),
-      ignore: ['mapping/*', 'markdown/*', 'exhentai/*'],
-    },
-  ];
-
-  const documentFiles = [];
-  mappings.map(({ id, category }) => {
-    const ext = category === 'mapping' ? 'json' : 'md';
-    const src = `src/assets/${category}/${id}.${ext}`;
-    const dist = `dist/${category}/${id}/${id}.${ext}`;
-    const distEditor = `dist/${category}-editor/${id}/${id}.${ext}`;
-    documentFiles.push({
-      from: handleWithPrefix(src),
-      to: handleWithPrefix(dist),
-    });
-    documentFiles.push({
-      from: handleWithPrefix(src),
-      to: handleWithPrefix(distEditor),
-    });
-  });
-
-  return [...assetsFiles, ...documentFiles];
-};
+const path = require('path');
+const { compressJSON, getCopyPluginProps, getHtmlPluginProps, getEntry } = require(path.join(__dirname, './utils'));
 
 const commonHtmlWebpackProps = {
-  template: handleWithPrefix('src/index.html'),
   minify: {
     minifyJS: true,
     minifyCSS: true,
@@ -73,85 +43,13 @@ const commonHtmlWebpackProps = {
   `,
 };
 
-const getHtmlPluginProps = mappings => {
-  const mainPageProps = [
-    new HtmlWebpackPlugin({
-      ...commonHtmlWebpackProps,
-      filename: 'index.html',
-      chunks: ['ninoninoni'],
-      title: `${author}'s ${name}`,
-      description: `${author}'s ${name}`,
-    }),
-    new HtmlWebpackPlugin({
-      ...commonHtmlWebpackProps,
-      filename: 'stock-shipment/index.html',
-      chunks: ['stock-shipment'],
-      title: `${author}'s business`,
-      description: `${author}'s business`,
-    }),
-  ];
-  const detailPageProps = [];
-  const editorPageProps = [];
-
-  mappings.map(({ id, category, title, type, subType }) => {
-    const commonTemplateProps = {
-      title: `${type} - ${subType} - ${title}`,
-      description: `${type} - ${subType}`,
-    };
-    switch (category) {
-      case 'mapping':
-        detailPageProps.push(
-          new HtmlWebpackPlugin({
-            ...commonHtmlWebpackProps,
-            ...commonTemplateProps,
-            filename: `${category}/${id}/index.html`,
-            chunks: ['mapping-detail'],
-          }),
-        );
-        editorPageProps.push(
-          new HtmlWebpackPlugin({
-            ...commonHtmlWebpackProps,
-            ...commonTemplateProps,
-            filename: `mapping-editor/${id}/index.html`,
-            chunks: ['mapping-editor'],
-          }),
-        );
-        break;
-
-      case 'markdown':
-        detailPageProps.push(
-          new HtmlWebpackPlugin({
-            ...commonHtmlWebpackProps,
-            ...commonTemplateProps,
-            filename: `${category}/${id}/index.html`,
-            chunks: ['markdown-detail'],
-          }),
-        );
-        editorPageProps.push(
-          new HtmlWebpackPlugin({
-            ...commonHtmlWebpackProps,
-            ...commonTemplateProps,
-            filename: `markdown-editor/${id}/index.html`,
-            chunks: ['markdown-editor'],
-          }),
-        );
-        break;
-
-      default:
-        break;
-    }
-  });
-
-  return [...mainPageProps, ...detailPageProps, ...editorPageProps];
-};
-
-const readableMappingFiles = initMappingFiles();
-const copyPluginProps = getCopyPluginProps(readableMappingFiles);
-const htmlPluginProps = getHtmlPluginProps(readableMappingFiles);
+// Don't merge these plugins into utils, for customed plugins
+const copyPluginProps = getCopyPluginProps();
+const htmlPluginProps = getHtmlPluginProps(commonHtmlWebpackProps);
 
 const plugins = [
   ...htmlPluginProps,
-  new CopyWebpackPlugin([...copyPluginProps]),
+  new CopyWebpackPlugin(copyPluginProps),
   new MomentLocalesPlugin({
     localesToKeep: ['zh-cn'],
   }),
@@ -161,13 +59,6 @@ const plugins = [
 ];
 
 module.exports = {
-  entry: {
-    'markdown-detail': handleWithPrefix('src/router/MarkdownDetailDataController.tsx'),
-    'markdown-editor': handleWithPrefix('src/router/MarkdownEditorDataController.tsx'),
-    'mapping-detail': handleWithPrefix('src/router/MappingDetailDataController.tsx'),
-    'mapping-editor': handleWithPrefix('src/router/MappingDetailDataController.tsx'),
-    'stock-shipment': handleWithPrefix('src/router/StockAndShipmentDataController.tsx'),
-    ninoninoni: handleWithPrefix('src'),
-  },
+  entry: getEntry(),
   plugins,
 };
